split all the file converters into their own files
